restructured the project files
