+++ src/app/app-routing.module.ts
@@ -4,20 +4,23 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomePageComponent} from './home-page/home-page.component';
 import { DescriptionTwitterComponent } from './home-page/description-twitter.component';
 
-import { MainComponent } from './main.component';
-import { UserComponent } from './side-panel/user.component';
-import { InfoComponent } from './side-panel/info.component';
+import { MainComponent } from './tool-main-panel/main.component';
+import { UserComponent } from './tool-side-panel/user.component';
+import { InfoComponent } from './tool-side-panel/info.component';
+import { HowItWorksComponent } from './home-page/how-it-works.component';
+import { CodeOfConductComponent } from './home-page/code-of-conduct.component';
 
 import { PageNotFoundComponent } from './page-not-found.component';
 
 const routes: Routes = [
 	{ path: 'homepage', component: HomePageComponent},
 	{ path: '', redirectTo: 'homepage', pathMatch: 'full'},
+    { path: 'howitworks', component: HowItWorksComponent},
+    { path: 'codeofconduct', component: CodeOfConductComponent},
     { path: 'twitter/:pid', component: MainComponent, children: [
     	{path: 'details/:uid', component: UserComponent},
         {path: '**', component: InfoComponent}
     ]},
-    { path: 'vk/:uid/:pid', component: MainComponent },
 	{ path: '**', component: PageNotFoundComponent }
 ];
