+++ src/app/app.module.ts
@@ -6,25 +6,27 @@ import { FormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
-import { MainComponent } from './main.component';
-import { HeaderComponent } from './header.component';
-import { UserComponent } from './side-panel/user.component';
-import { UserTimelineComponent } from './side-panel/user.timeline.component';
-import { InfoComponent } from './side-panel/info.component';
-import { D3Component } from './d3.component';
-import { SearchComponent } from './search.component';
-import { TweetComponent } from './tweet.component';
+import { MainComponent } from './tool-main-panel/main.component';
+import { HeaderComponent } from './multi-use/header.component';
+import { UserComponent } from './tool-side-panel/user.component';
+import { UserTimelineComponent } from './tool-side-panel/user.timeline.component';
+import { InfoComponent } from './tool-side-panel/info.component';
+import { D3Component } from './tool-main-panel/d3.component';
+import { SearchComponent } from './multi-use/search.component';
+import { TweetComponent } from './multi-use/tweet.component';
 import { PageNotFoundComponent } from './page-not-found.component';
-import { LoadingComponent } from './loading.component';
-import { ShareComponent } from './share.component';
-import { HelpComponent } from './help.component';
+import { LoadingComponent } from './multi-use/loading.component';
+import { ShareComponent } from './tool-main-panel/share.component';
+import { HelpComponent } from './tool-main-panel/help.component';
+import { HowItWorksComponent } from './home-page/how-it-works.component';
+import { CodeOfConductComponent } from './home-page/code-of-conduct.component';
 
 import { HomePageComponent } from './home-page/home-page.component';
 import { DescriptionTwitterComponent } from './home-page/description-twitter.component';
 
-import { ApiService } from './api.service';
-import { MainCommunicationService } from './main.communication.service';
-import { HighlightService } from './highlight.service';
+import { ApiService } from './services/api.service';
+import { MainCommunicationService } from './services/main.communication.service';
+import { HighlightService } from './services/highlight.service';
 
 @NgModule({
   declarations: [
@@ -42,7 +44,9 @@ import { HighlightService } from './highlight.service';
     ShareComponent,
     HelpComponent,
     HomePageComponent,
-      DescriptionTwitterComponent,
+    DescriptionTwitterComponent,
+    HowItWorksComponent,
+    CodeOfConductComponent
   ],
   imports: [
     BrowserModule,
