+++ src/app/home-page/code-of-conduct.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'code-of-conduct',
+    templateUrl: './code-of-conduct.component.html',
+    styleUrls: ['./code-of-conduct.component.scss']
+})
+
+export class CodeOfConductComponent {
+
+}
