+++ src/app/home-page/how-it-works.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'how-it-works',
+    templateUrl: './how-it-works.component.html',
+    styleUrls: ['./how-it-works.component.scss']
+})
+
+export class HowItWorksComponent {
+    
+}
