+++ src/app/multi-use/header.component.html
@@ -3,14 +3,17 @@
     <div class="logo">
         <img 
             alt="the tracemap logo"
-            src="./../assets/images/fav64.png">
+            src="/assets/images/fav64.png">
         <p>TraceMap</p>
     </div>
     <div class="header-links">
-        <button class="about-us">
+        <button>
             ABOUT<br>US
         </button>
-        <button class="code-of-conduct">
+        <button>
+            HOW IT<br>WORKS
+        </button>
+        <button>
             CODE OF<br>CONDUCT
         </button>
     </div>
