+++ src/app/multi-use/header.component.scss
@@ -50,7 +50,7 @@
             font-weight: 600;
             line-height: 17px;
             border-radius: 5px;             
-            margin-right: 10px;
+            margin-right: 8px;
 
             &:hover {
                 color: #000;
