+++ src/app/multi-use/search.component.scss
@@ -37,7 +37,7 @@ input[type=text] {
 
 button {
   grid-area: button;
-  background: url("./../assets/images/loupe.svg") no-repeat center transparent;
+  background: url("/assets/images/loupe.svg") no-repeat center transparent;
   background-size: contain;
   border: 3px solid transparent;
   outline: none;
