+++ src/app/services/api.service.ts
@@ -3,7 +3,7 @@ import { Http, Response } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
-import { environment } from './../environments/environment';
+import { environment } from './../../environments/environment';
 
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/mergeMap';
