+++ src/app/services/highlight.service.ts
