+++ src/app/tool-main-panel/d3.component.scss
