+++ src/app/tool-main-panel/help.component.scss
