+++ src/app/tool-main-panel/main.component.scss
