+++ src/app/tool-main-panel/share.component.scss
