+++ src/app/tool-main-panel/share.component.ts
