+++ src/app/tool-side-panel/info.component.scss
