+++ src/app/tool-side-panel/user.component.scss
